servers/server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
subgrounds = "^1.7.1"
jupyter = "^1.0.0"
ipykernel = "^6.25.2"
web3 = "^6.10.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
